requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
Pillow==10.1.0
weasyprint==60.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
